+++ common/src/commonMain/kotlin/foatto/core/model/response/chart/ChartElementLineData.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 class ChartElementLineData(
     val x: Int,
     var y: Float,
-    val color: ULong,
+    val color: Int,
 )
