+++ common/src/commonMain/kotlin/foatto/core/model/response/chart/ChartElementTextData.kt
@@ -6,8 +6,8 @@ import kotlinx.serialization.Serializable
 class ChartElementTextData(
     val x1: Int,
     val x2: Int,
-    val fillColor: ULong,
-    val borderColor: ULong,
-    val textColor: ULong,
+    val fillColor: Int,
+    val borderColor: Int,
+    val textColor: Int,
     val text: String,
 )
