+++ common/src/commonMain/kotlin/foatto/core/model/response/chart/ChartLegendData.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 class ChartLegendData(
-    val fillColor: ULong?,
-    val borderColor: ULong,
-    val textColor: ULong,
+    val fillColor: Int?,
+    val borderColor: Int,
+    val textColor: Int,
     val text: String,
 )
